+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Query string value type
+- `HasEnvVars` to determine if the `EnvVarSet` has variables defined
 
 
 ## [0.4.0] - 2018-06-12